Create JavaRegex.java
This one needs some explanation so here's the directions 

"Write a class called myRegex which will contain a string pattern. You need to write a regular expression and assign it to the pattern such that it can be used to validate an IP address. Use the following definition of an IP address:"